* master: (58 commits)
  VIC-12632: Update lib/audio to latest victor_master (#5165)
  VIC-13576 Remove BehaviorProxGetToDistance (#5137)
  VIC-9888 README.md - Explain VICTOR_COMPAT_VERSION
  VIC-11532 Python SDK: nightly tests fix #5168
  Tech Debt: Fix line break error in vic-crashuploader-init script (#5146)
  VIC-13583: Lowered priority of ReactToUncalibratedHeadAndLift  (#5160)
  Update version 1.6.0 (#5166)
  Bruce/VIC-13004-sdk-play-audio (#5152)
  VIC-11532 Python SDK nightly tests: turn off unreliable tests #5164
  makes logging helper (#5162)
  just use proxData.foundObject in habitat detection (#5156)
  VIC-13609 Fix issues with saving channel console vars (#5155)
  VIC-9888 Add /anki/etc/victor-compat-version
  VIC-2243: Forward kernel panic logs to backtrace (#5130)
  VIC-5814: Retire special case ERRORS_TO_IGNORE (#5159)
  VIC-13382: Add support for measuring and graphing vic-robot latency
  Updating victor-animations-assets to v4916 (#5157)
  Held-In-Palm Bug Fixes (#5123)
  VIC-12822 SDK sleep seizure (#5147)
  BI-1850 Retry VM provisioning
  ...

+++ coretech/vision/shared/spriteSequence/spriteSequence.h
@@ -13,6 +13,7 @@
 #define __Vision_Shared_SpriteSequence_H__
 
 #include "coretech/vision/shared/spriteCache/spriteCache.h"
+#include "coretech/common/shared/types.h"
 
 #include <string>
 #include <deque>
@@ -42,7 +43,7 @@ public:
   SpriteSequence(LoopConfig config = LoopConfig::Hold);
   virtual ~SpriteSequence();
   
-  uint GetNumFrames() const { return static_cast<uint>(_frames.size()); }
+  uint16_t GetNumFrames() const { return static_cast<uint16_t>(_frames.size()); }
   
   void AddFrame(Vision::SpriteHandle spriteHandle);
